value-pairs: move ValuePairs struct declaration to a private header
To make the ValuePairs struct opaque to users, move its definition to
a private header, so that cmdline.c can have access, but still the
rest of the codebase can't.
